Use size_t for sizes (arity, no. variables, other things that follow)
This should suppress all warnings from -Wconversion (but -Wno-sign-conversion).
Note that I tried to avoid casts, so I was rather liberal in changing types.
"Necessary" casts are highlighted either as "safe casts" (when I was sure not
to break things), or as "dubious casts" when I was not so sure.

+++ src/minizip/unzip.c
@@ -1583,7 +1583,7 @@ extern ZPOS64_T ZEXPORT unzGetCurrentFileZStreamPos64(unzFile file) {
   return <0 with error code if there is an error
     (UNZ_ERRNO for IO error, or zLib error for uncompress error)
 */
-extern int ZEXPORT unzReadCurrentFile(unzFile file, voidp buf, unsigned len) {
+extern int ZEXPORT unzReadCurrentFile(unzFile file, voidp buf, size_t len) {
     int err=UNZ_OK;
     uInt iRead = 0;
     unz64_s* s;
